# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Electronic Educare API",
    "description": "API documentation for the Electronic Educare backend.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/assignment/fetch": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/assignment/add": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "class": {
                    "example": "any"
                  },
                  "marks": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "dueDate": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/assignment/submit": {
      "post": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assignmentId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/assignment/submissions": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/assignment/my-submissions": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/fetch": {
      "get": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add": {
      "post": {
        "tags": [
          "Subjects"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "class": {
                    "example": "any"
                  },
                  "marks": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "dueDate": {
                    "example": "any"
                  },
                  "courseName": {
                    "example": "any"
                  },
                  "courseCode": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  },
                  "credits": {
                    "example": "any"
                  },
                  "department": {
                    "example": "any"
                  },
                  "prerequisites": {
                    "example": "any"
                  },
                  "instructor": {
                    "example": "any"
                  },
                  "startingDate": {
                    "example": "any"
                  },
                  "venue": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "time": {
                    "example": "any"
                  },
                  "noOfStudents": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "schoolName": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "consent": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "rating": {
                    "example": "any"
                  },
                  "feedback": {
                    "example": "any"
                  },
                  "schoolId": {
                    "example": "any"
                  },
                  "stream": {
                    "example": "any"
                  },
                  "fees": {
                    "example": "any"
                  },
                  "programType": {
                    "example": "any"
                  },
                  "programLabel": {
                    "example": "any"
                  },
                  "totalStudents": {
                    "example": "any"
                  },
                  "installments": {
                    "example": "any"
                  },
                  "desc": {
                    "example": "any"
                  },
                  "subjectName": {
                    "example": "any"
                  },
                  "code": {
                    "example": "any"
                  },
                  "classId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/submit": {
      "post": {
        "tags": [
          "Behaviour"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assignmentId": {
                    "example": "any"
                  },
                  "studentClass": {
                    "example": "any"
                  },
                  "subject": {
                    "example": "any"
                  },
                  "questionType": {
                    "example": "any"
                  },
                  "startTime": {
                    "example": "any"
                  },
                  "endTime": {
                    "example": "any"
                  },
                  "correct": {
                    "example": "any"
                  },
                  "incorrect": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/submissions": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/my-submissions": {
      "get": {
        "tags": [
          "Assignments"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}